+++ bin/sidewalk-pave.py
@@ -3,7 +3,7 @@
 import optparse
 import traceback
 
-import sidewalk
+import sidewalk.loggers
 import sidewalk.manager
 import sidewalk.exceptions
 from sidewalk.activity_aggregator import ActivityAggregator
@@ -47,28 +47,33 @@ def main():
 		
 		aggregator.run()
 	except sidewalk.exceptions.SidewalkSettingsFileIOError, e:
-		sidewalk.error_log('Settings file IOError for requested file "%s"' % (
+		sidewalk.loggers.error('Settings file IOError for requested file "%s"' % (
 			e.filename
 		))
+	except sidewalk.exceptions.SidewalkSectionNotDefined, e:
+		sidewalk.loggers.error('Settings file "%s" does not have the required section "%s"' % (
+			e.filename,
+			e.section
+		))
 	except sidewalk.exceptions.SidewalkKeyDoesNotExist, e:
-		sidewalk.error_log('Invalid key "%s"' % (
+		sidewalk.loggers.error('Invalid key "%s"' % (
 			e.key
 		))
 	except sidewalk.exceptions.SidewalkGroupDoesNotExist, e:
-		sidewalk.error_log('Invalid group key "%s"' % (
+		sidewalk.loggers.error('Invalid group key "%s"' % (
 			e.group_key
 		))
 	except sidewalk.exceptions.SidewalkModuleImportError, e:
-		sidewalk.error_log('Cannot import module "%s"' % (
+		sidewalk.loggers.error('Cannot import module "%s"' % (
 			e.module
 		))
 	except sidewalk.exceptions.SidewalkMethodDoesNotExist, e:
-		sidewalk.error_log('Method "%s" not defined in module "%s"' % (
+		sidewalk.loggers.error('Method "%s" not defined in module "%s"' % (
 			e.method,
 			e.module
 		))
 	except sidewalk.exceptions.SidewalkRogueActivityProcessor, e:
-		sidewalk.error_log('Activity processor "%s" threw an unhandled exception' % (
+		sidewalk.loggers.error('Activity processor "%s" threw an unhandled exception' % (
 			e.activity_processor
 		))