+++ sidewalk_activity_processors/example.py
@@ -4,16 +4,18 @@ sidewalk_activity_processors.example
 This module contains sample activity processors.
 """
 
+import sidewalk.loggers
+
 def hello():
 	"""Simply print hello world back to the console."""
 	
-	print 'hello world'
+	sidewalk.loggers.log('hello world')
 
 def import_test():
 	"""Import the core sidewalk module to check against a defined value."""
 	
 	import sidewalk
-	print 'sidewalk == %s' % (sidewalk.__title__)
+	sidewalk.loggers.debug('sidewalk == %s' % (sidewalk.__title__))
 
 def baddie_unhandled_test():
 	"""A test activity processor that throws an uncaught exception."""